+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Note the the latest version is usually work in progress and may have not yet bee
 ## Changed
 
 - `--sqlite-dsn` -> `--db-path`. Reads from `ENVELOPE_DB_PATH` env var now too
+- made all tests parallel
 
 # v0.0.3